Finished part4 exercises

+++ controllers/blogs.js
@@ -1,35 +1,29 @@
 const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
     response.json(blogs)
   } catch(error) {
     next(error)
   }
 })
 
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')){
-    return authorization.substring(7)
-  }
-  return null
-}
 
 blogsRouter.post('', async (request, response, next) => {
   try {
     const body = request.body
-    const token = getTokenFrom(request)
 
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if(!token || !decodedToken.id){
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!request.token || !decodedToken.id){
       return response.status(401).json({ error: 'Token missing or invalid' })
     }
 
-    const user = await user.findById(decodedToken.id)
+    //console.log(decodedToken)
+    const user = await User.findById(decodedToken.id)
 
     const blog = new Blog({
       title: body.title,
@@ -41,7 +35,8 @@ blogsRouter.post('', async (request, response, next) => {
 
     const savedBlog = await blog.save()
 
-    user.notes = user.notes.concat(savedBlog._id)
+    //console.log(user)
+    user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
 
     response.status(201).json(savedBlog)
@@ -53,6 +48,18 @@ blogsRouter.post('', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+
+    const blog = await Blog.findById(request.params.id)
+    //console.log(blog)
+    const user = await User.findById(blog.user)
+    //console.log(user)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!request.token || !decodedToken.id || !(decodedToken.id === user.id)){
+      return response.status(401).json({ error: 'Token missing or invalid' })
+    }
+
+    //console.log(decodedToken)
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch(error) {
